+++ R/rNodal-package.r
@@ -5,3 +5,20 @@
 #' @name rNodal
 #' @docType package
 NULL
+
+
+#' dataset from Guo, example calculations
+#'
+#' @usage data(dpt)
+#' @keywords datasets
+#' @docType data
+"dpt"
+
+
+#' Digitized version of Fancher and Brown curve. Brown Chapter 2 figure 41
+#'
+#' @usage data(curves.0241)
+#' @keywords datasets
+#' @docType data
+"curves.0241"
+